+++ src/main/java/fll/scheduler/ScheduleWriter.java
@@ -551,7 +551,7 @@ public final class ScheduleWriter {
       // check if team is missing an opponent
       if (null == schedule.findOpponent(si, performance)) {
         teamsMissingOpponents.add(si);
-        row.setAttribute("background-color", "magenta");
+        row.setAttribute("background-color", "lightgrey");
       }
 
       final LocalTime performanceTime = performance.getTime();
@@ -637,8 +637,6 @@ public final class ScheduleWriter {
       for (final TeamScheduleInfo si : teamsMissingOpponents) {
         final Element stayingTableRow = FOPUtils.createTableRow(document);
         stayingTableBody.appendChild(stayingTableRow);
-
-        stayingTableRow.setAttribute("background-color", "magenta");
 
         final Element stayingTableCell = FOPUtils.createTableCell(document, FOPUtils.TEXT_ALIGN_CENTER,
                                                                   String.format(formatString, si.getTeamNumber()));