+++ src/main/java/fll/web/api/JudgesServlet.java
@@ -55,7 +55,8 @@ public class JudgesServlet extends HttpServlet {
     final HttpSession session = request.getSession();
     final AuthenticationContext auth = SessionAttributes.getAuthentication(session);
 
-    if (!auth.isJudge()) {
+    if (!auth.isJudge()
+        && !auth.isReportGenerator()) {
       response.sendError(HttpServletResponse.SC_FORBIDDEN);
       return;
     }