+++ src/main/java/fll/web/api/NumericCategoryScoresServlet.java
@@ -51,7 +51,8 @@ public class NumericCategoryScoresServlet extends HttpServlet {
     final AuthenticationContext auth = SessionAttributes.getAuthentication(session);
 
     if (!auth.isRef()
-        && !auth.isJudge()) {
+        && !auth.isJudge()
+        && !auth.isReportGenerator()) {
       response.sendError(HttpServletResponse.SC_FORBIDDEN);
       return;
     }