+++ src/main/java/fll/web/api/PlayoffBracketTeamsServlet.java
@@ -47,7 +47,8 @@ public class PlayoffBracketTeamsServlet extends HttpServlet {
     final AuthenticationContext auth = SessionAttributes.getAuthentication(session);
 
     if (!auth.isJudge()
-        && !auth.isRef()) {
+        && !auth.isRef()
+        && !auth.isReportGenerator()) {
       response.sendError(HttpServletResponse.SC_FORBIDDEN);
       return;
     }