+++ src/main/java/fll/web/report/AddAwardWinner.java
@@ -146,7 +146,8 @@ public class AddAwardWinner extends BaseFLLServlet {
                                 final HttpSession session)
       throws IOException, ServletException {
     final AuthenticationContext auth = SessionAttributes.getAuthentication(session);
-    if (!auth.requireRoles(request, response, session, Set.of(UserRole.JUDGE, UserRole.HEAD_JUDGE), false)) {
+    if (!auth.requireRoles(request, response, session,
+                           Set.of(UserRole.JUDGE, UserRole.HEAD_JUDGE, UserRole.REPORT_GENERATOR), false)) {
       return;
     }