+++ src/main/java/fll/web/report/ScaledSubjectiveByAwardGroup.java
@@ -66,7 +66,7 @@ public class ScaledSubjectiveByAwardGroup extends BaseFLLServlet {
 
     final AuthenticationContext auth = SessionAttributes.getAuthentication(session);
 
-    if (!auth.requireRoles(request, response, session, Set.of(UserRole.JUDGE), false)) {
+    if (!auth.requireRoles(request, response, session, Set.of(UserRole.JUDGE, UserRole.REPORT_GENERATOR), false)) {
       return;
     }